backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-cors==4.0.0